+++ CHANGELOG.md
@@ -22,6 +22,8 @@ all APIs might be changed.
 - Querygen no longer duplicates variable names in generated `ArgumentStruct`s
   when an argument is used twice in a query.
 - Fixed support for `Float` scalars in the generator.
+- Fixed support for fields named `self`, `super` etc. which can't be made into
+  raw identifiers.
 
 ## v0.15.0 - 2021-09-23